+++ two_factor/views/utils.py
@@ -1,6 +1,12 @@
+import logging
+from django.contrib.formtools.wizard.forms import ManagementForm
+from django.core.exceptions import ValidationError
 from django.utils.decorators import method_decorator
 from django.utils.functional import lazy_property
+from django.utils.translation import ugettext as _
 from two_factor.compat import SessionStorage, SessionWizardView
+
+logger = logging.getLogger(__name__)
 
 
 class ExtraSessionStorage(SessionStorage):
@@ -79,6 +85,40 @@ class IdempotentSessionWizardView(SessionWizardView):
         conditions have changed.
         """
         if self.steps.current not in self.steps.all:
+            logger.warning("Current step '%s' is no longer valid, returning "
+                           "to last valid step in the wizard.",
+                           self.steps.current)
+            return self.render_goto_step(self.steps.all[-1])
+
+        # -- Duplicated code from upstream
+        # Look for a wizard_goto_step element in the posted data which
+        # contains a valid step name. If one was found, render the requested
+        # form. (This makes stepping back a lot easier).
+        wizard_goto_step = self.request.POST.get('wizard_goto_step', None)
+        if wizard_goto_step and wizard_goto_step in self.get_form_list():
+            return self.render_goto_step(wizard_goto_step)
+
+        # Check if form was refreshed
+        management_form = ManagementForm(self.request.POST, prefix=self.prefix)
+        if not management_form.is_valid():
+            raise ValidationError(
+                _('ManagementForm data is missing or has been tampered.'),
+                code='missing_management_form',
+            )
+
+        form_current_step = management_form.cleaned_data['current_step']
+        if (form_current_step != self.steps.current and
+                self.storage.current_step is not None):
+            # form refreshed, change current step
+            self.storage.current_step = form_current_step
+        # -- End duplicated code from upstream
+
+        # This is different from the first check, as this checks
+        # if the new step is available. See issue #65.
+        if self.steps.current not in self.steps.all:
+            logger.warning("Requested step '%s' is no longer valid, returning "
+                           "to last valid step in the wizard.",
+                           self.steps.current)
             return self.render_goto_step(self.steps.all[-1])
 
         return super(IdempotentSessionWizardView, self).post(*args, **kwargs)